# Optimize Color/Brush parsing and string conversion, reduce allocs