--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/ColorKind.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/ColorKind.cs
@@ -5,7 +5,7 @@
 namespace System.Windows.Media
 {
     /// <summary>
-    /// Encapsulates color kinds as categorized via <see cref="KnownColors.MatchColor(string)"/>
+    /// Encapsulates color kinds as categorized via <see cref="KnownColors.MatchColor(ReadOnlySpan{char})"/>
     /// </summary>
     internal enum ColorKind
     {