--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/Parsers.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/Parsers.cs
@@ -195,7 +195,8 @@ namespace MS.Internal
             bool isNumericColor;
             bool isScRgbColor;
             bool isContextColor;
-            string trimmedColor = KnownColors.MatchColor(color, out isPossibleKnowColor, out isNumericColor, out isContextColor, out isScRgbColor);
+            string trimmedColor = color.Trim();
+            KnownColors.MatchColor(trimmedColor, out isPossibleKnowColor, out isNumericColor, out isContextColor, out isScRgbColor);
 
             if ((isPossibleKnowColor == false) &&
                 (isNumericColor == false) &&
@@ -243,7 +244,8 @@ namespace MS.Internal
             bool isNumericColor;
             bool isScRgbColor;
             bool isContextColor;
-            string trimmedColor = KnownColors.MatchColor(brush, out isPossibleKnownColor, out isNumericColor, out isContextColor, out isScRgbColor);
+            string trimmedColor = brush.Trim();
+            KnownColors.MatchColor(trimmedColor, out isPossibleKnownColor, out isNumericColor, out isContextColor, out isScRgbColor);
 
             if (trimmedColor.Length == 0)
             {