… src/Microsoft.DotNet.Wpf/src/Shared/MS/Internal/ValueTokenizerHelper.cs