@h3xds1nz 2025-03-24
Resolved merge conflicts, validated with #10633 and #10632 tests.